kvm: Load kernel at 0x1000
As described in Documentation/x86/boot.txt, memory starting at 0x1000 is
available for the boot loader.  As we don't have a separate boot loader,
load the kernel at the address. Note: we're setting up the segment
selector here so 0x0100:0x0000 in 16-bit real mode is 0x1000 in physical
memory.

Signed-off-by: Zhiyuan Dai <jerry.zhiyuan.dai@gmail.com>