kvm, bios: Get rid of buggy int10 handler
At moment any BIOS code using which needs own stack have to be remade.
It's a bit not that easy task. We're happily may live without int10
handler since we switch to early_printk interface really early. And
actually we don't scan VGA memory anyway so that even if there is
something written we ignore the result.

Signed-off-by: Zhiyuan Dai <jerry.zhiyuan.dai@gmail.com>