kvm: Fix kernel loading error messages
This patch fixes kernel loading error messages. We expect users to try
to load bzImage most of the time and. However, if the user passes a
binary that's not a bzImage, lets just print out the following warning:

Warning: vmlinux is not a bzImage. Trying to load it as a flat binary...

and pass the control to the binary.

Signed-off-by: Zhiyuan Dai <jerry.zhiyuan.dai@gmail.com>